rename swagger file
