+++ internal/frontend/dags/convert.go
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-package dag
+package dags
 
 import (
 	"github.com/daguflow/dagu/internal/dag"
